Misc changes for code review fix

+++ sdk/eventhub/azure-eventhubs/azure/eventhub/client_abstract.py
@@ -171,7 +171,7 @@ class EventHubClientAbstract(object):
         :type http_proxy: dict[str, Any]
         :param auth_timeout: The time in seconds to wait for a token to be authorized by the service.
          The default value is 60 seconds. If set to 0, no timeout will be enforced from the client.
-        :type auth_timeout: int
+        :type auth_timeout: float
         :param user_agent: The user agent that needs to be appended to the built in user agent string.
         :type user_agent: str
         :param max_retries: The max number of attempts to redo the failed operation when an error happened. Default
@@ -188,10 +188,10 @@ class EventHubClientAbstract(object):
         :type max_batch_size: int
         :param receive_timeout: The timeout time in seconds to receive a batch of events from an Event Hub.
          Default value is 0 seconds.
-        :type receive_timeout: int
+        :type receive_timeout: float
         :param send_timeout: The timeout in seconds for an individual event to be sent from the time that it is
          queued. Default value is 60 seconds. If set to 0, there will be no timeout.
-        :type send_timeout: int
+        :type send_timeout: float
 
         Example:
             .. literalinclude:: ../examples/test_examples_eventhub.py
@@ -227,7 +227,7 @@ class EventHubClientAbstract(object):
         :type http_proxy: dict[str, Any]
         :param auth_timeout: The time in seconds to wait for a token to be authorized by the service.
          The default value is 60 seconds. If set to 0, no timeout will be enforced from the client.
-        :type auth_timeout: int
+        :type auth_timeout: float
         :param user_agent: The user agent that needs to be appended to the built in user agent string.
         :type user_agent: str
         :param max_retries: The max number of attempts to redo the failed operation when an error happened. Default
@@ -244,10 +244,10 @@ class EventHubClientAbstract(object):
         :type max_batch_size: int
         :param receive_timeout: The timeout time in seconds to receive a batch of events from an Event Hub.
          Default value is 0 seconds.
-        :type receive_timeout: int
+        :type receive_timeout: float
         :param send_timeout: The timeout in seconds for an individual event to be sent from the time that it is
          queued. Default value is 60 seconds. If set to 0, there will be no timeout.
-        :type send_timeout: int
+        :type send_timeout: float
 
         Example:
             .. literalinclude:: ../examples/test_examples_eventhub.py
